requirements.txt
Flask==1.1.2
gunicorn==20.0.4
google-cloud-speech==1.2.0
google-cloud-storage==1.37.1
srt==3.0.0
